feat: solve the modules task
